Normalize symlinks
