+++ crates/biome_plugin_loader/src/lib.rs
@@ -14,9 +14,9 @@ use std::sync::Arc;
 use biome_analyze::{AnalyzerPlugin, AnalyzerPluginVec};
 use biome_console::markup;
 use biome_deserialize::json::deserialize_from_json_str;
-use biome_fs::FileSystem;
+use biome_fs::{FileSystem, normalize_path};
 use biome_json_parser::JsonParserOptions;
-use camino::{Utf8Component, Utf8Path, Utf8PathBuf};
+use camino::{Utf8Path, Utf8PathBuf};
 use plugin_manifest::PluginManifest;
 
 #[derive(Debug)]
@@ -91,39 +91,6 @@ impl BiomePlugin {
 
         Ok(plugin)
     }
-}
-
-/// Normalizes the given `path` without requiring filesystem access.
-///
-/// This only normalizes `.` and `..` entries, but does not resolve symlinks.
-fn normalize_path(path: &Utf8Path) -> Utf8PathBuf {
-    let mut stack = Vec::new();
-
-    for component in path.components() {
-        match component {
-            Utf8Component::ParentDir => {
-                if stack.last().is_some_and(|last| *last == "..") {
-                    stack.push("..");
-                } else {
-                    stack.pop();
-                }
-            }
-            Utf8Component::CurDir => {}
-            Utf8Component::RootDir => {
-                stack.clear();
-                stack.push("/");
-            }
-            Utf8Component::Normal(c) => stack.push(c),
-            _ => {}
-        }
-    }
-
-    let mut result = Utf8PathBuf::new();
-    for part in stack {
-        result.push(part);
-    }
-
-    result
 }
 
 #[cfg(test)]
